sellshoes/pyproject.toml
[tool.poetry]
name = "sellshoes"
version = "0.1.0"
description = ""
authors = ["believeohiozua <believe.ohiozua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
django = "5.1.4"
python-decouple = "3.8"
sqlparse = "0.5.3"
django-cors-headers = "4.6.0"
djangorestframework = "3.15.2"
drf-yasg = "1.21.8"
django-filter = "^24.3"
cloudinary = "1.41.0"
dj-database-url = "^2.3.0"
psycopg = "^3.2.3"
pytest = "8.3.4"
requests = "2.32.3"
certifi = "2024.12.14"
charset-normalizer = "3.4.1"
django-cloudinary-storage = "0.3.0"
djangorestframework-simplejwt = "5.3.1"
idna = "3.10"
inflection = "0.5.1"
iniconfig = "2.0.0"
packaging = "24.2"
pluggy = "1.5.0"
pyjwt = "2.10.1"
pytest-django = "4.9.0"
pytz = "2024.2"
pyyaml = "6.0.2"
six = "1.17.0"
uritemplate = "4.1.1"
urllib3 = "2.3.0"
schemathesis = "^3.39.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
